requirements.txt
opencv-python
numpy
Pillow
matplotlib
scipy
tk
